+++ src/main/java/org/stellar/sdk/AssetTypeCreditAlphaNum.java
@@ -1,6 +1,5 @@
 package org.stellar.sdk;
 
-import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NonNull;
@@ -12,7 +11,6 @@ import lombok.NonNull;
  *     href="https://developers.stellar.org/docs/learn/fundamentals/stellar-data-structures/assets"
  *     target="_blank">Assets</a>
  */
-@AllArgsConstructor
 @EqualsAndHashCode(callSuper = false)
 @Getter
 public abstract class AssetTypeCreditAlphaNum extends Asset {
@@ -21,6 +19,23 @@ public abstract class AssetTypeCreditAlphaNum extends Asset {
 
   /** Asset issuer */
   @NonNull protected final String issuer;
+
+  /**
+   * Class constructor
+   *
+   * @param code Asset code
+   * @param issuer Asset issuer
+   * @throws IllegalArgumentException when code is invalid or issuer is not a valid Ed25519 public
+   *     key
+   */
+  public AssetTypeCreditAlphaNum(@NonNull String code, @NonNull String issuer) {
+    if (!StrKey.isValidEd25519PublicKey(issuer)) {
+      throw new IllegalArgumentException("Invalid issuer: " + issuer);
+    }
+
+    this.code = code;
+    this.issuer = issuer;
+  }
 
   @Override
   public String toString() {